更新至0.4.0(新增世界模式曲目获取支持)
1.新增世界模式曲目获取支持
2.修复了Legacy章节体力消耗固定为2的bug
3.修复了Legacy章节残片深化不起作用的bug
4.新增独立的 "/user/me" API接口
5.修复了无法自定义世界模式满体力数的bug(现可在fixed_properties表的max_stamina中设置大于0且可不等于12的满体力数)

+++ Team123it.Arcaea.MarveCube/Processors/Front/Purchase.cs
@@ -267,7 +267,7 @@ namespace Team123it.Arcaea.MarveCube.Processors.Front
 								if (World2.CalculateCurrentStaminas(worldTimeFullRecharged.AddMinutes(-6 * 30), out bool isOverflowStamina) == World2.FullStaminas)
 								{
 									worldTimeFullRecharged = worldTimeFullRecharged.AddMinutes(-6 * 30);
-									totalStaminas = World2.FullStaminas;
+									totalStaminas = (int)World2.FullStaminas;
 									if (isOverflowStamina)
 									{
 										int addOverflowStaminas = Convert.ToInt32(Math.Floor((DateTime.Now - worldTimeFullRecharged).TotalMinutes / 30));
@@ -313,7 +313,7 @@ namespace Team123it.Arcaea.MarveCube.Processors.Front
 									if (World2.CalculateCurrentStaminas(worldTimeFullRecharged.AddMinutes(-6 * 30), out bool isOverflowStamina) == World2.FullStaminas)
 									{
 										worldTimeFullRecharged = worldTimeFullRecharged.AddMinutes(-6 * 30);
-										totalStaminas = World2.FullStaminas;
+										totalStaminas = (int)World2.FullStaminas;
 										if (isOverflowStamina)
 										{
 											int addOverflowStaminas = Convert.ToInt32(Math.Floor((DateTime.Now - worldTimeFullRecharged).TotalMinutes / 30));
